Fixed pivot of Sally and Annie, Completed Sally and Annie prefab, Fixed read/write of knife, Fixed portal always showing, fixed ending always triggering, Adjusted player prefab
