+++ Assets/Prefabs/Entities/Humans/V2/Annie.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f92a2f55739deac458785ef15a1d9ea2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
