+++ Assets/Prefabs/Entities/Humans/V2/Sally.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dc92a6ffed116c4409db607ee76702cf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
