+++ Assets/Scripts/Misc/CrimeSceneStateHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 775dbc97303a43b4d8a6ac8b009e68b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
